Remove `not_standard_first_version` check
This check shows this message:

```
Also, note the warning: This looks like a new registration that registers version 0.1.1.
Ideally, you should register an initial release with 0.0.1, 0.1.0 or 1.0.0 version numbers
This can be safely ignored. However, if you want to fix this you can do so. Call
register() again after making the fix. This will update the Pull request.
```

This is confusing (it prompts users to change version number even if they do not
have to) and false: it is not true that changing the version number and
registering again will update the pull request, on the contrary it will open a
new pull request which will never be merged because it will conflict with the
previously opened PR.