Add option to push to a fork of the registry repo (#68)
Adds a `registry_fork` keyword argument to the `register` method. Registration commits will be pushed to a branch on the forked registry. `registry_fork` defaults to `registry` if it is unspecified.

This is useful when you don't want the bot/person using RegistryTools.jl to have write access to the actual Registry.